% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rateable.R
\name{rateable}
\alias{rateable}
\title{This function takes a vector of postcodes and returns the rateable values,
split by under/over 51k, within these postcodes.}
\usage{
rateable(pcd, return = "raw")
}
\arguments{
\item{pcd}{A vector of postcodes (character) of form xxxx xxx/ xxx xxx /...}

\item{return}{defaults to 'raw' - returns the raw data for the postcode list, also takes
'scatSummary' - provides the sum of rateable values by SCat,
'primarySummary' - provides the sum of rateable values by primary code}
}
\description{
This function takes a vector of postcodes and returns the rateable values,
split by under/over 51k, within these postcodes.
}
\examples{
rateable(c('SW15 1SF'),return=scatSummary) 

}
\keyword{chris}
